chore(keyboard): move default event listeners to a separate file
