+++ lib/features/keyboard/Keyboard.js
@@ -3,6 +3,12 @@ import {
   matches as domMatches
 } from 'min-dom';
 
+import {
+  hasModifier,
+  isCmd,
+  isShift
+} from './helpers';
+
 
 /**
  * A keyboard abstraction that may be activated and
@@ -24,14 +30,12 @@ import {
  *
  * @param {Config} config
  * @param {EventBus} eventBus
- * @param {EditorActions} editorActions
  */
-export default function Keyboard(config, eventBus, editorActions) {
+export default function Keyboard(config, eventBus) {
   var self = this;
 
   this._config = config || {};
   this._eventBus = eventBus;
-  this._editorActions = editorActions;
 
   this._listeners = [];
 
@@ -81,14 +85,11 @@ export default function Keyboard(config, eventBus, editorActions) {
   eventBus.on('detach', function() {
     self.unbind();
   });
-
-  this._init();
 }
 
 Keyboard.$inject = [
   'config.keyboard',
-  'eventBus',
-  'editorActions'
+  'eventBus'
 ];
 
 
@@ -125,171 +126,6 @@ Keyboard.prototype._fire = function(event) {
   this._eventBus.fire('keyboard.' + event, { node: this._node, listeners: this._listeners });
 };
 
-Keyboard.prototype._init = function() {
-
-  var listeners = this._listeners;
-
-  var editorActions = this._editorActions,
-      config = this._config;
-
-  // init default listeners
-
-  // undo
-  // (CTRL|CMD) + Z
-  function undo(key, modifiers) {
-
-    if (isCmd(modifiers) && !isShift(modifiers) && key === 90) {
-      editorActions.trigger('undo');
-
-      return true;
-    }
-  }
-
-  // redo
-  // CTRL + Y
-  // CMD + SHIFT + Z
-  function redo(key, modifiers) {
-
-    if (isCmd(modifiers) && (key === 89 || (key === 90 && isShift(modifiers)))) {
-      editorActions.trigger('redo');
-
-      return true;
-    }
-  }
-
-  // copy
-  // CTRL/CMD + C
-  function copy(key, modifiers) {
-
-    if (isCmd(modifiers) && (key === 67)) {
-      editorActions.trigger('copy');
-
-      return true;
-    }
-  }
-
-  // paste
-  // CTRL/CMD + V
-  function paste(key, modifiers) {
-
-    if (isCmd(modifiers) && (key === 86)) {
-      editorActions.trigger('paste');
-
-      return true;
-    }
-  }
-
-  /**
-   * zoom in one step
-   * CTRL + +
-   *
-   * 107 = numpad plus
-   * 187 = regular plus
-   * 171 = regular plus in Firefox (german keyboard layout)
-   *  61 = regular plus in Firefox (US keyboard layout)
-   */
-  function zoomIn(key, modifiers) {
-
-    if ((key === 107 || key === 187 || key === 171 || key === 61) && isCmd(modifiers)) {
-      editorActions.trigger('stepZoom', { value: 1 });
-
-      return true;
-    }
-  }
-
-  /**
-   * zoom out one step
-   * CTRL + -
-   *
-   * 109 = numpad minus
-   * 189 = regular minus
-   * 173 = regular minus in Firefox (US and german keyboard layout)
-   */
-  function zoomOut(key, modifiers) {
-
-    if ((key === 109 || key === 189 || key === 173) && isCmd(modifiers)) {
-      editorActions.trigger('stepZoom', { value: -1 });
-
-      return true;
-    }
-  }
-
-  /**
-   * zoom to the default level
-   * CTRL + 0
-   *
-   * 96 = numpad zero
-   * 48 = regular zero
-   */
-  function zoomDefault(key, modifiers) {
-
-    if ((key === 96 || key === 48) && isCmd(modifiers)) {
-      editorActions.trigger('zoom', { value: 1 });
-
-      return true;
-    }
-  }
-
-  // delete selected element
-  // DEL
-  function removeSelection(key, modifiers) {
-
-    if (key === 46) {
-      editorActions.trigger('removeSelection');
-
-      return true;
-    }
-  }
-
-  // move canvas left
-  // left arrow
-  //
-  // 37 = Left
-  // 38 = Up
-  // 39 = Right
-  // 40 = Down
-  function moveCanvas(key, modifiers) {
-
-    if ([37, 38, 39, 40].indexOf(key) >= 0) {
-
-      var opts = {
-        invertY: config.invertY,
-        speed: (config.speed || 50)
-      };
-
-      switch (key) {
-      case 37: // Left
-        opts.direction = 'left';
-        break;
-      case 38: // Up
-        opts.direction = 'up';
-        break;
-      case 39: // Right
-        opts.direction = 'right';
-        break;
-      case 40: // Down
-        opts.direction = 'down';
-        break;
-      }
-
-      editorActions.trigger('moveCanvas', opts);
-
-      return true;
-    }
-  }
-
-  listeners.push(undo);
-  listeners.push(redo);
-  listeners.push(copy);
-  listeners.push(paste);
-  listeners.push(removeSelection);
-  listeners.push(zoomIn);
-  listeners.push(zoomOut);
-  listeners.push(zoomDefault);
-  listeners.push(moveCanvas);
-};
-
-
 /**
  * Add a listener function that is notified with (key, modifiers) whenever
  * the keyboard is bound and the user presses a key.
@@ -303,22 +139,3 @@ Keyboard.prototype.addListener = function(listenerFn) {
 Keyboard.prototype.hasModifier = hasModifier;
 Keyboard.prototype.isCmd = isCmd;
 Keyboard.prototype.isShift = isShift;
-
-
-function hasModifier(modifiers) {
-  return (modifiers.ctrlKey || modifiers.metaKey || modifiers.shiftKey || modifiers.altKey);
-}
-
-function isCmd(modifiers) {
-  // ensure we don't react to AltGr
-  // (mapped to CTRL + ALT)
-  if (modifiers.altKey) {
-    return false;
-  }
-
-  return modifiers.ctrlKey || modifiers.metaKey;
-}
-
-function isShift(modifiers) {
-  return modifiers.shiftKey;
-}
