+++ lib/features/keyboard/helpers/hasModifier.js
@@ -1,0 +1,7 @@
+/**
+ * Returns true if event was triggered with any modifier
+ * @param {KeyboardEvent} event
+ */
+export function hasModifier(event) {
+  return (event.ctrlKey || event.metaKey || event.shiftKey || event.altKey);
+}
