+++ lib/features/keyboard/helpers/isCmd.js
@@ -1,0 +1,12 @@
+/**
+ * @param {KeyboardEvent} event
+ */
+export function isCmd(event) {
+  // ensure we don't react to AltGr
+  // (mapped to CTRL + ALT)
+  if (event.altKey) {
+    return false;
+  }
+
+  return event.ctrlKey || event.metaKey;
+}
