+++ lib/features/keyboard/helpers/isShift.js
@@ -1,0 +1,6 @@
+/**
+ * @param {KeyboardEvent} event
+ */
+export function isShift(event) {
+  return event.shiftKey;
+}
