+++ lib/features/keyboard/index.js
@@ -1,6 +1,11 @@
+import EditorActionsModule from 'lib/features/editor-actions';
+
 import Keyboard from './Keyboard';
+import KeyboardBindings from './KeyboardBindings';
 
 export default {
-  __init__: [ 'keyboard' ],
-  keyboard: [ 'type', Keyboard ]
+  __depends__: [ EditorActionsModule ],
+  __init__: [ 'keyboard', 'keyboardBindings' ],
+  keyboard: [ 'type', Keyboard ],
+  keyboardBindings: [ 'type', KeyboardBindings ]
 };
